% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardLibrary_030117.R
\docType{data}
\name{StandardLibrary_030117}
\alias{StandardLibrary_030117}
\title{Example standard library}
\format{A dataframe with 298 rows and 14 columns}
\description{
Example standard library for use in identifying metabolites. Can be used as argument for standardLibrary flag in ConsensusAlign function
}
\author{
Ryne Ramaker <ryneramaker@gmail.com>
}
\keyword{data}
